refactored linkedwikis and media from entities to types folder
