+++ src/Database/Entities/types/ILinkedWikis.ts
