+++ src/Database/Entities/wiki.entity.ts
@@ -20,11 +20,11 @@ import Tag from './tag.entity'
 import User from './user.entity'
 import Language from './language.entity'
 import Metadata from './metadata.entity'
-import Media from './media.entity'
+import Media from './types/IMedia'
 import Image from './image.entity'
 import { Author } from './types/IUser'
 import dateMiddleware from './middlewares/wikiMiddleware'
-import LinkedWikis from './linkedWikis.entity'
+import LinkedWikis from './types/ILinkedWikis'
 
 @ObjectType()
 @Entity()
